Split time series into seperate page of demonstrator
